feat: show loading indicator when timeline is empty
Show a `ContainedLoadingIndicator` in the center of the screen when the timeline is empty.
This provides a visual cue to the user that the app is loading data and there are no timeline items to display yet.

Previously, only the `PullRefreshIndicator` was shown, which could be confusing to the user.