feat: add splash screen
Added a splash screen to the app using the SplashScreen API. This screen is displayed while the app is loading and is automatically dismissed when loading is complete.

The splash screen uses the app's icon and a background color defined in the theme. It is also set to be kept on screen until the token loading is complete.