refactor: Update GetMyFavoriteUseCase to return only notes
The `GetMyFavoriteUseCase` has been updated to return only the note data from the API response, instead of the entire response object. This change simplifies the data structure and makes it easier to use in the UI layer.

Additionally, the `AccountRepository` has been updated to return a list of `IFavoritesResponse` objects, which contain the note data along with other metadata.

A new data class `IFavoritesResponse` has been added to represent the API response. This class includes the note data as well as the creation timestamp, ID, and note ID.