refactor: update GetMyFavoriteUseCase to return UiState
Update `GetMyFavoriteUseCase` to return `FavoritesScreenUiState` instead of a list of `NotesTimelineResponse`.
This change allows the presenter to directly update the UI state with the result of the use case, simplifying the code and improving the separation of concerns.

Additionally, error handling is added to the use case to ensure that the UI is updated appropriately in case of an exception.