refactor: move note create event handling to NoteCreatePresenter
Moved the logic for handling note creation events from HomeScreenPresenter and NotificationScreenPresenter to a dedicated NoteCreatePresenter. This centralizes the logic and improves code organization.
Added isSuccessCreateNote property to the HomeScreen and NotificationScreen's State classes for access to the note creation status, and removed it from NotificationUiState and HomeScreenUiState classes.
Added NoteCreatePresenter, NoteCreateState and NoteCreateEvent for the new logic.